Snapshot? idk, let's see if this works
Tried to put a hero section on the homepage, had a couple problems with my IDE, so I'm not sure if this will work as I expect it to.
If it does, however, I will put the changes on the other pages too.

I still need to upload the STEMS to Petrichor... -_-